+++ frontend/wailsjs/go/app/App.js
@@ -42,8 +42,8 @@ export function ClearOfflineBodyCache(arg1) {
   return window['go']['app']['App']['ClearOfflineBodyCache'](arg1);
 }
 
-export function Contacts_BrowseContacts(arg1, arg2, arg3, arg4) {
-  return window['go']['app']['App']['Contacts_BrowseContacts'](arg1, arg2, arg3, arg4);
+export function Contacts_BrowseContacts(arg1, arg2, arg3, arg4, arg5) {
+  return window['go']['app']['App']['Contacts_BrowseContacts'](arg1, arg2, arg3, arg4, arg5);
 }
 
 export function Contacts_CreateContact(arg1) {